# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem Customs Clearance Bea Cukai (Simulasi Jalur Merah \u0026 Jalur Hijau).",
        "title": "Customs Clearance API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v1",
    "paths": {
        "/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menampilkan statistik pengajuan barang, kategori risiko, top komoditas, top pelabuhan, dan transaksi terbaru (Khusus Officer)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard / Seeder"
                ],
                "summary": "Ringkasan Dashboard Bea Cukai",
                "responses": {
                    "200": {
                        "description": "Dashboard Berhasil Diambil",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.DashboardSummary"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Gagal Mengambil Data",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/seed": {
            "post": {
                "description": "Membersihkan seluruh database (TRUNCATE CASCADE) dan mengisi data tiruan Bea Cukai dalam jumlah besar untuk keperluan pengetesan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard / Seeder"
                ],
                "summary": "Seed Data Tiruan",
                "responses": {
                    "200": {
                        "description": "Seeding Berhasil",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal Melakukan Seeding",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.Clearance": {
            "type": "object",
            "properties": {
                "commodity": {
                    "$ref": "#/definitions/models.Commodity"
                },
                "commodity_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "port": {
                    "$ref": "#/definitions/models.Port"
                },
                "port_id": {
                    "type": "integer"
                },
                "risk_profile": {
                    "$ref": "#/definitions/models.RiskProfile"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                },
                "valuation": {
                    "type": "number"
                }
            }
        },
        "models.Commodity": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "hs_code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "import_duty_rate": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "vat_rate": {
                    "type": "number"
                }
            }
        },
...
        "models.Country": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.DashboardSummary": {
            "type": "object",
            "properties": {
                "recent_clearances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Clearance"
                    }
                },
                "risk_counts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "status_counts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "top_commodities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TopCommodity"
                    }
                },
                "top_ports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TopPort"
                    }
                },
                "total_clearances": {
                    "type": "integer"
                },
                "total_valuation": {
                    "type": "number"
                }
            }
        },
...
        "models.Port": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "country": {
                    "$ref": "#/definitions/models.Country"
                },
                "country_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.RiskProfile": {
            "type": "object",
            "properties": {
                "clearance_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "description": "HIGH or LOW",
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.TopCommodity": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "hs_code": {
                    "type": "string"
                },
                "total_value": {
                    "type": "number"
                }
            }
        },
        "models.TopPort": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "total_value": {
                    "type": "number"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Tipe token: Bearer \u003cJWT_TOKEN\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}